+++ tales_django/entities/Tracks/__init__.py
@@ -1,9 +1,11 @@
-from .admin import TrackAdmin
-from .models import Track
+from .admin import SeriesAdmin, TrackAdmin
+from .models import Series, Track
 
 # from .forms import TrackAdminForm
 
 __all__ = [
+    'Series',
+    'SeriesAdmin',
     'Track',
     'TrackAdmin',
     # 'TrackAdminForm',