+++ tales_django/entities/Tracks/admin/__init__.py
@@ -1,5 +1,6 @@
 from .AuthorAdmin import AuthorAdmin
 from .RubricAdmin import RubricAdmin
+from .SeriesAdmin import SeriesAdmin
 from .TagAdmin import TagAdmin
 from .TrackAdmin import TrackAdmin
 from .UserTrackAdmin import UserTrackAdmin
@@ -7,6 +8,7 @@ from .UserTrackAdmin import UserTrackAdmin
 __all__ = [
     'AuthorAdmin',
     'RubricAdmin',
+    'SeriesAdmin',
     'TagAdmin',
     'TrackAdmin',
     'UserTrackAdmin',