+++ tales_django/entities/Tracks/forms/__init__.py
@@ -1,0 +1,2 @@
+from .TrackAdminForm import TrackAdminForm
+from .TrackInlineForm import TrackInlineForm