+++ tales_django/entities/Tracks/views/__init__.py
@@ -1,4 +1,6 @@
-from .author_details_view import author_details_view
+# Import views only when needed to avoid circular imports
+# from .author_details_view import author_details_view
+
 from .author_index_view import author_index_view
 from .author_sitemap import author_sitemap
 from .favorites_view import favorites_view
@@ -9,13 +11,13 @@ from .tag_index_view import tag_index_view
 from .track_details_view import track_details_view
 
 __all__ = [
-    'track_details_view',
-    'author_sitemap',
-    'author_index_view',
     'author_details_view',
-    'rubric_index_view',
-    'rubric_details_view',
-    'tag_index_view',
-    'tag_details_view',
+    'author_index_view',
+    'author_sitemap',
     'favorites_view',
+    'rubric_details_view',
+    'rubric_index_view',
+    'tag_details_view',
+    'tag_index_view',
+    'track_details_view',
 ]